Rename Typings => Types and fix eslint
