+++ .eslintrc.json
@@ -48,6 +48,7 @@
           "",
           ".js",
           ".ts",
+          ".d.ts",
           ".tsx",
           ".es.ts",
           ".tsx",
