+++ Types/funbox.d.ts
